// apps/server/src/events/events.types.ts
export enum EventVariant {
  INFO = 'INFO',
  WARNING = 'WARNING',
  NEGATIVE = 'NEGATIVE',
  POSITIVE = 'POSITIVE',
}
// ...
export enum EventDataActionType {
  CREATE = 'CREATE',
  UPDATE = 'UPDATE',
  ARCHIVED = 'ARCHIVED',
}

registerEnumType(EventDataActionType, {
  name: 'EventDataActionType',
  description: 'Available event data action types',
});

registerEnumType(EventDataActionType, {
  name: 'EventDataActionType',
  description: 'Available event data action types',
});

export enum EventChannel {
  WORKSPACE = 'WORKSPACE',
  PERSONAL = 'PERSONAL',
  NONE = 'NONE',
}

registerEnumType(EventChannel, {
  name: 'EventChannel',
  description: 'Available event channels',
});

export class CaptureEventInput {
  @ApiProperty()
  @IsString()
  @IsOptional()
  workspaceId?: string;

  @ApiProperty()
  @IsString()
  @IsOptional()
  sessionId?: string;

  @ApiProperty()
  @IsEnum(EventChannel)
  @IsOptional()
  channel?: EventChannel;

  @ApiProperty()
  @IsEnum(EventType)
  type: EventType;

  @ApiProperty()
  @IsOptional()
  @IsEnum(EventDataActionType)
  actionType?: EventDataActionType;

  @ApiProperty()
  @IsOptional()
  @IsEnum(EventVariant)
  variant?: EventVariant;

  @ApiProperty()
  @IsString()
  @IsOptional()
  userId?: string;

  @ApiProperty()
  @IsString()
  @IsOptional()
  ref?: string;

  @ApiProperty()
  @IsObject()
  @IsOptional()
  data?: any;

  @ApiProperty()
  @IsBoolean()
  @IsOptional()
  persist?: boolean;

  @ApiProperty()
  @IsNumber()
  @IsOptional()
  time?: number;

  @ApiProperty()
  @IsObject()
  @IsOptional()
  relatedEntities?: RelatedEntity[];

  @ApiProperty()
  @Type(() => ReportTimeSeriesInput)
  @ValidateNested()
  @IsOptional()
  reportTimeRange?: ReportTimeSeriesInput;
}

export interface SocketAuthInput {
  token: string;
  deviceId: string;
}

export interface SocketJoinWorkspaceInput {
  workspaceId: string;
  token: string;
  deviceId: string;
}

export interface SocketListenUsersInput {
  addUserIds?: string[];
  removeUserIds?: string[];
}

export interface SendUserEventInput {
  userId: string;
  eventName: string;
  data?: any;
}

// ======================= SDK Related =======================
export interface SdkSocketJoinDto {
  key: string;
}
export interface SdkSendEventDto {
  workspaceId: string;
  type: EventType;
  data: EventEntity;
}